Fix compiler warnings about unused formal parameters, unused initialized variables, name hiding, sign mismatches, and narrowing.

+++ test/async_pipe.cpp
@@ -69,7 +69,7 @@ BOOST_AUTO_TEST_CASE(multithreaded_async_pipe)
     asio::io_context ioc;
 
     std::vector<std::thread> threads;
-    for (int i = 0; i < std::thread::hardware_concurrency(); i++)
+    for (size_t i = 0; i < std::thread::hardware_concurrency(); i++)
     {
         threads.emplace_back([&ioc]
         {
